requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
tensorflow
keras
streamlit
Pillow
joblib